feat: 필수 파라미터 NULL일 때 에러코드 추가(#67)

+++ fourtune/src/main/java/com/fourtune/auction/boundedContext/watchList/application/service/WatchListSupport.java
@@ -6,6 +6,8 @@ import com.fourtune.auction.boundedContext.watchList.domain.WatchListUser;
 import com.fourtune.auction.boundedContext.watchList.port.out.WatchListItemsRepository;
 import com.fourtune.auction.boundedContext.watchList.port.out.WatchListRepository;
 import com.fourtune.auction.boundedContext.watchList.port.out.WatchListUserRepository;
+import com.fourtune.auction.global.error.ErrorCode;
+import com.fourtune.auction.global.error.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -20,10 +22,16 @@ public class WatchListSupport {
     private final WatchListRepository watchListRepository;
 
     public boolean existsByUserIdAndAuctionItemId(Long userId, Long auctionId){
+        if(userId == null || auctionId == null)
+            throw new BusinessException(ErrorCode.MISSING_INPUT_VALUE);
+
         return watchListRepository.existsByUserIdAndAuctionItemId(userId, auctionId);
     }
 
     public void deleteByUserIdAndAuctionItemId(Long userId, Long auctionId){
+        if(userId == null || auctionId == null)
+            throw new BusinessException(ErrorCode.MISSING_INPUT_VALUE);
+
         watchListRepository.deleteByUserIdAndAuctionItemId(userId, auctionId);
     }
